Crate liberdus_rpc

This is the main module of the JSON-RPC server.

This module sets up and initializes the server, managing shared state, cryptographic utilities, archiver discovery, and routing for both HTTP and WebSocket-based RPC calls.

The server leverages tokio for asynchronous operations and poem for http. RPC is custom implemented on top of poem.

Features

Loads configuration and seeds archivers from a predefined file.

Supports HTTP and WebSocket endpoints for handling JSON-RPC requests.

Automatically discovers active archivers and updates nodelists periodically.

Employs a shared state for thread-safe interaction with the liberdus backend.

Includes CORS middleware for cross-origin requests.

Modules

rpc: Implementation of the JSON-RPC protocol for HTTP and WebSocket.

methods: Implements the RPC methods invoked by JSON-RPC requests.

config: Manages configuration loading and validation.

archivers: Includes utilities for managing and discovering archivers.

liberdus: This module includes load aware node selectioin mechnism and chat room subscription manager.

crypto: Provides shardus cryptographic utilities.

collector: Utilities for quering chain data from distribution protocol.

Entry Point

The main function initializes the server, manages shared state, and launches the application.

Modules

archivers 🔒

Archiver Utility Module

collector 🔒

This module provides utilities to interact with a transaction collector API.

config 🔒

Configuration for the rpc server.

crypto 🔒

liberdus 🔒

This module contains the node management logic require for load balancing with consensor nodes

methods 🔒

RPC Methods for Interacting with the Liberdus Backend.

rpc 🔒

This module provides utilities to handle RPC calls for interacting with a backend system.

Structs

CrossThreadSharedState

Shared state accessible across multiple threads.

Functions

main 🔒

The main entry point for the application.